HELP ME JESUS CHRIST

+++ Kernel/include/lib.h
@@ -13,9 +13,9 @@ void * memcpy(void * destination, const void * source, uint64_t length);
 char *cpuVendor(char *result);
 int myStrcmp(const char s1, const char s2);
 //freezes for the amount of ms indicated
-void sleepms(int ms);
+void sleepms(uint32_t ms);
 void reverseStr(char* str, int length);
-void intToStr(int num, char* str);
+void intToStr(uint64_t num, char* str);
 void printMm();
 void runTestMm();
 void testeando();
